dora-hardware/vendors/imu/MTI-710/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
adafruit-blinka=8.20.0=pypi_0
adafruit-circuitpython-busdevice=5.2.6=pypi_0
adafruit-circuitpython-register=1.9.16=pypi_0
adafruit-circuitpython-requests=2.0.1=pypi_0
adafruit-circuitpython-typing=1.9.4=pypi_0
adafruit-platformdetect=3.47.0=pypi_0
adafruit-pureio=1.1.11=pypi_0
alabaster=0.7.13=pyhd8ed1ab_0
astroid=2.5.8=py37h89c1867_0
babel=2.12.1=pyhd8ed1ab_1
brotli-python=1.0.9=py37hd23a5d3_7
ca-certificates=2023.5.7=hbcca054_0
certifi=2022.12.7=py37h06a4308_0
charset-normalizer=3.2.0=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
docutils=0.16=py37h89c1867_3
dora-rs=0.2.3=pypi_0
evdev=1.6.1=pypi_0
idna=3.4=pyhd8ed1ab_0
imagesize=1.4.1=pyhd8ed1ab_0
isort=5.6.2=py37hc8dfbb8_0
jinja2=2.11.3=pyhd8ed1ab_2
lazy-object-proxy=1.6.0=py37h27cfd23_0
libedit=3.1.20221030=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markupsafe=1.1.1=py37h5e8e339_3
mccabe=0.6.1=py_1
ncurses=6.4=h6a678d5_0
openssl=1.0.2u=h7b6447c_0
packaging=23.1=pyhd8ed1ab_0
pip=22.3.1=py37h06a4308_0
py-imu-mpu6050=0.3=pypi_0
pyftdi=0.54.0=pypi_0
pygments=2.15.1=pyhd8ed1ab_0
pylint=2.7.2=py37h89c1867_0
pynput=1.7.6=pypi_0
pyserial=3.5=pypi_0
pysocks=1.7.1=py37h89c1867_5
python=3.7.0=h6e4f718_3
python-xlib=0.33=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytz=2023.3=pyhd8ed1ab_0
pyusb=1.2.1=pypi_0
readline=7.0=h7b6447c_5
requests=2.31.0=pyhd8ed1ab_0
setuptools=65.6.3=py37h06a4308_0
setuptools-lint=0.6.0=pyh9f0ad1d_0
six=1.16.0=pyh6c4a22f_0
smbus=1.1.post2=pypi_0
snowballstemmer=2.2.0=pyhd8ed1ab_0
sphinx=3.5.4=pyh44b312d_0
sphinxcontrib-applehelp=1.0.4=pyhd8ed1ab_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=2.0.1=pyhd8ed1ab_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.5=pyhd8ed1ab_2
sqlite=3.33.0=h62c20be_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
typed-ast=1.4.3=py37h7f8727e_1
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
urllib3=2.0.4=pyhd8ed1ab_0
wheel=0.38.4=py37h06a4308_0
wrapt=1.11.2=py37h8f50634_1
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
